modularized the sessionController
